+++ packages/generator/src/edmx-to-vdm/v2/function-import.ts
@@ -26,9 +26,8 @@ export function generateFunctionImportsV2(
         httpMethod,
         serviceMetadata.swagger
       );
-
       // TODO 1571 remove when supporting entity type as parameter
-      if (hasUnsupportedParameterTypes(f.Name, f.Parameter)) {
+      if (hasUnsupportedParameterTypes(f)) {
         return undefined;
       }